+++ src/app/[locale]/page.tsx
@@ -25,7 +25,7 @@ function ViewHome(props: { locale: string }) {
 }
 
 function ViewCity(props: { city: string; handlers: Handlers }) {
-  return <CityPage key="city" city={props.city} bordells={bordells} onBackAction={props.handlers.onBackHome} onBordellClickAction={props.handlers.onBordellClick} />
+  return <CityPage key="city" city={props.city} bordells={bordells} onBackAction={props.handlers.onBackHome} />
 }
 
 function ViewDetail(props: { bordell: Bordell; hasCity: boolean; handlers: Handlers }) {